refactor : RunningExitConfirmationDialog를 싱글에서도 재사용할 수 있도록 위치 변경 #PAR-336
